# openapi.yaml
openapi: 3.0.1
info:
  title: Linux Command Plugin
  description: A plugin that allows the user to run terminal commands.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /run_cmd/{command}:
    get:
      operationId: runCommand
      summary: Run Linux Command
      parameters:
      - in: path
        name: command
        schema:
            type: string
        required: true
        description: The linux command to run.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getrunResponse'

  /get_available_commands/:
    get:
      operationId: getAvailableCommands
      summary: Get available non default commands
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAvailableCommandsResponse'

  /install_cmd/{package}:
    get:
      operationId: installCommand
      summary: Install Ubuntu package using sudo
      parameters:
      - in: path
        name: package
        schema:
            type: string
        required: true
        description: The linux command that install a package
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/installCommandResponse'

components:
  schemas:
    getrunResponse:
      type: object
      properties:
        commands:
          type: array
          items:
            type: string
          description: The output of the linux command ran
    getAvailableCommandsResponse:
      type: object
      properties:
        commands:
          type: array
          items:
            type: string
          description: The list of available non default commands.
    installCommandResponse:
      type: object
      properties:
        message:
          type: string
          description: A message indicating whether the package was successfully installed.